Move attachment_drag_and_drop and markdown_leading_tag_handler into subdirectories
Reorganize:
- attachment_drag_and_drop.js → attachments/drag_and_drop.js
- markdown_leading_tag_handler.js → markdown/leading_tag_handler.js
