+++ src/editor/attachments/drag_and_drop.js
