+++ src/editor/markdown/leading_tag_handler.js
